Remove submodule update
Motivation
----------
In order to run the validation jenkins job with certain SDKs, the .git folder has to be removed. This prevents cloning the submodule for couchbase-net-client, however the Dockerfile for building the performer has now been modified to perform a fresh clone of the SDK instead of using the submodule.